✨ feat: implement user migration and update product model

+++ src/domain/entities/product.go
@@ -6,6 +6,6 @@ type Product struct {
 	ID          uuid.UUID
 	Name        string
 	Description string
-	Price       float64
+	Price       uint
 	Stock       uint
 }
